+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Files unexpectedly existing in ~/.config/pext/modules/ no longer causes a crash
+
 ## [0.4.1] - 2017-03-05
 ### Changed
 - The default window is no longer explicitly borderless